Add files via upload
The new addition to the code gives the corn a "grade" based on the number of pixels in a "bin". Each bin represents a 16pts with 1 being the least yellow and 8 being the most yellow. 

To do: write to CSV/append as iterating through the code and comment on each code block above in markdown